Sentiment_Analyzer_v1
This upload is the folder containing code for the sentiment analyzer application. At the moment there is only one view in use. However, three templates have been created. Note the application connects with postgresql instead of sqlite. This is a development version only. Although the settings.py file contains a dummy user name and password to access a postgresql database, no other user names or passwords are included.